+++ lib/api/boot/assets/check.png
