+++ lib/api/boot/assets/favicon.ico
