+++ lib/api/boot/assets/logo-mobile.png
