+++ lib/api/boot/assets/logo.png
