+++ lib/api/boot/index.js
@@ -139,7 +139,7 @@ app.use('/api', require('lib/api/whitelist-api'))
  * Comment API Service
  */
 
-app.use('/api', require('lib/site/comment'))
+app.use('/api', require('lib/api/comment-api'))
 
 /**
  * Forums API Service
@@ -197,5 +197,5 @@ app.use(require('lib/site/help'))
 app.use(require('lib/site/homepage'))
 app.use(require('lib/site/topic'))
 app.use(require('lib/site/notifications-page'))
-app.use(require('lib/site/forum'))
+app.use(require('lib/site/forum-new'))
 app.use(require('lib/404'))
