+++ lib/forum/not-allowed/not-allowed.js
@@ -1,4 +1,4 @@
-import render from '../../../render/render'
+import render from '../../render/render'
 import template from './template.jade'
 
 export function verifyForumAccess (ctx, next) {
