+++ lib/server/index.js
@@ -4,7 +4,7 @@
 
 var async = require('async')
 var config = require('lib/config')
-var app = require('lib/boot')
+var app = require('lib/api/boot')
 var serverFactory = require('lib/server-factory')
 var notifier = require('democracyos-notifier')
 var notifications = require('lib/notifications').notifier
