+++ lib/settings/boot/boot.js
@@ -23,6 +23,7 @@ let log = debug('democraycos:boot')
  * Mount applications.
  */
 
+import '../../forum/forum'
 import '../../analytics'
 import '../../body-classes/body'
 import '../../content-lock/locker'
@@ -42,6 +43,13 @@ timeago('.ago', { lang: config.locale, interval: 1000 })
 /**
  * Reload the page when it's not handled by the front end
  */
+page('/settings/', function (ctx) {
+  window.location.reload(false)
+})
+
+page('/settings/*', function (ctx) {
+  window.location.reload(false)
+})
 
 page('*', function (ctx) {
   log('Should reload.')
