+++ lib/site/boot/boot.js
@@ -20,6 +20,20 @@ import '../../debug'
 let log = debug('democraycos:boot')
 
 /**
+ * Reload the settings page
+ */
+
+page('/settings/*', function (ctx) {
+  log('Should reload.')
+  window.location.reload(false)
+})
+
+page('/settings/', function (ctx) {
+  log('Should reload.')
+  window.location.reload(false)
+})
+
+/**
  * Mount applications.
  */
 
@@ -35,7 +49,8 @@ import '../../site/forgot/forgot'
 import '../../site/notifications-page/notifications'
 import '../../logout/logout'
 import '../../site/auth-facebook/auth-facebook'
-import '../../site/forum/forum'
+import '../../forum/forum'
+import '../../site/forum-new/forum-new'
 import '../../site/homepage/homepage'
 import '../../site/newsfeed/newsfeed'
 import '../../site/topic/topic'
